Format code and optimize imports
