+++ common/src/main/java/com/microsoft/identity/common/internal/servertelemetry/ServerTelemetry.java
@@ -5,8 +5,6 @@ import android.content.Context;
 import com.microsoft.identity.common.internal.cache.ISharedPreferencesFileManager;
 import com.microsoft.identity.common.internal.cache.SharedPreferencesFileManager;
 import com.microsoft.identity.common.internal.logging.Logger;
-import com.microsoft.identity.common.internal.providers.oauth2.AuthorizationResult;
-import com.microsoft.identity.common.internal.providers.oauth2.TokenResult;
 import com.microsoft.identity.common.internal.result.AcquireTokenResult;
 
 import java.util.HashMap;
@@ -39,7 +37,7 @@ public class ServerTelemetry {
     }
 
     public static void emit(Map<String, String> telemetry) {
-        for(Map.Entry<String, String> entry: telemetry.entrySet()) {
+        for (Map.Entry<String, String> entry : telemetry.entrySet()) {
             emit(entry.getKey(), entry.getValue());
         }
     }
@@ -130,8 +128,8 @@ public class ServerTelemetry {
 
         final ISharedPreferencesFileManager sharedPreferencesFileManager =
                 new SharedPreferencesFileManager(
-                    context,
-                    LAST_REQUEST_TELEMETRY_SHARED_PREFERENCES
+                        context,
+                        LAST_REQUEST_TELEMETRY_SHARED_PREFERENCES
                 );
 
         final IRequestTelemetryCache lastRequestTelemetryCache =
@@ -162,12 +160,12 @@ public class ServerTelemetry {
         sLastRequestTelemetry = new RequestTelemetry(sCurrentRequestTelemetry.getSchemaVersion(), false);
 
         // grab whatever common fields we can from current request
-        for(Map.Entry<String, String> entry: sCurrentRequestTelemetry.getCommonTelemetry().entrySet()) {
+        for (Map.Entry<String, String> entry : sCurrentRequestTelemetry.getCommonTelemetry().entrySet()) {
             putForLast(entry.getKey(), entry.getValue());
         }
 
         // grab whatever platform fields we can from current request
-        for(Map.Entry<String, String> entry: sCurrentRequestTelemetry.getPlatformTelemetry().entrySet()) {
+        for (Map.Entry<String, String> entry : sCurrentRequestTelemetry.getPlatformTelemetry().entrySet()) {
             putForLast(entry.getKey(), entry.getValue());
         }
     }
@@ -187,14 +185,14 @@ public class ServerTelemetry {
                     TAG + methodName,
                     "Last Request Telemetry Cache object was null. " +
                             "Unable to save request telemetry to cache."
-                    );
+            );
         }
 
         clearCurrentRequestTelemetry();
     }
 
     public static void completeScenario(String correlationId, AcquireTokenResult acquireTokenResult) {
-        final String errorCode = errorFromAcquireTokenResult(acquireTokenResult);
+        final String errorCode = TelemetryUtils.errorFromAcquireTokenResult(acquireTokenResult);
         completeScenario(correlationId, errorCode);
     }
 
@@ -211,31 +209,6 @@ public class ServerTelemetry {
             put(Schema.CURRENT_REQUEST_HEADER_NAME, getCurrentTelemetryHeaderString());
             put(Schema.LAST_REQUEST_HEADER_NAME, getLastTelemetryHeaderString());
         }};
-    }
-
-    public static String errorFromAcquireTokenResult(final AcquireTokenResult acquireTokenResult) {
-        final String errorFromAuthorization = getErrorFromAuthorizationResult(acquireTokenResult.getAuthorizationResult());
-        if (errorFromAuthorization != null) {
-            return errorFromAuthorization;
-        } else {
-            return getErrorFromTokenResult(acquireTokenResult.getTokenResult());
-        }
-    }
-
-    private static String getErrorFromAuthorizationResult(final AuthorizationResult authorizationResult) {
-        if (authorizationResult != null && authorizationResult.getErrorResponse() != null) {
-            return authorizationResult.getErrorResponse().getError();
-        } else {
-            return null;
-        }
-    }
-
-    private static String getErrorFromTokenResult(final TokenResult tokenResult) {
-        if (tokenResult != null && tokenResult.getErrorResponse() != null) {
-            return tokenResult.getErrorResponse().getError();
-        } else {
-            return null;
-        }
     }
 
 }
